myworld/travels/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15.0,<4.0
django-cors-headers>=4.3.1,<5.0
